+++ eyes_selenium/applitools/selenium/visual_grid/running_test.py
@@ -13,6 +13,8 @@ from .vg_task import VGTask
 if typing.TYPE_CHECKING:
     from typing import Any, Dict, List, Optional, Text
 
+    from structlog import BoundLogger
+
     from applitools.common import (
         RenderBrowserInfo,
         RenderStatusResults,
@@ -21,6 +23,7 @@ if typing.TYPE_CHECKING:
     )
     from applitools.common.selenium import Configuration
     from applitools.selenium.fluent import SeleniumCheckSettings
+    from applitools.selenium.visual_grid.rendreing_service import RenderingService
 
     from .eyes_connector import EyesConnector
 
@@ -57,6 +60,7 @@ class RunningTestCheck(object):
     region_selectors = attr.ib()  # type: List[List[VisualGridSelector]]
     source = attr.ib()  # type: Text
 
+    logger = attr.ib(init=False, repr=False)  # type: BoundLogger
     regions = attr.ib(init=False, factory=list)
     queue = attr.ib(init=False, factory=deque)
 
@@ -69,15 +73,16 @@ class RunningTestCheck(object):
         render_task = self._render_task(tag, short_description, render_request)
         self.queue = deque([render_task])
 
+    def __attrs_post_init__(self):
+        self.logger = self.running_test.logger.bind(running_test_check=self)
+
     def __hash__(self):
         return hash(self.name + self.uuid)
 
     def _render_task(self, tag, short_description, render_request):
         def render_task_succeeded(render_status):
             # type: (List[RenderStatusResults]) -> None
-            logger.debug(
-                "render_task_succeeded: task.uuid: {}".format(render_task.uuid)
-            )
+            self.logger.debug("render_task_succeeded", task_uuid=render_task.uuid)
             if render_status:
                 self.running_test.eyes.render_status_for_task(
                     render_task.uuid, render_status
@@ -85,13 +90,13 @@ class RunningTestCheck(object):
                 if render_status.status == RenderStatus.RENDERED:
                     for vgr in render_status.selector_regions:
                         if vgr.error:
-                            logger.error(vgr.error)
+                            self.logger.error("Region error", vgr_error=vgr.error)
                         else:
                             self.regions.append(vgr.to_region())
-                    logger.debug(
-                        "render_task_succeeded: uuid: {}\n\tregions {}".format(
-                            render_task.uuid, self.regions
-                        )
+                    self.logger.debug(
+                        "render_task_succeeded rendered",
+                        task_uuid=render_task.uuid,
+                        regions=self.regions,
                     )
                     # schedule check task
                     self.queue.append(self._check_task(render_task, tag))
@@ -106,16 +111,12 @@ class RunningTestCheck(object):
                     self.running_test.abort()
                     self.running_test.becomes_tested()
             else:
-                logger.error(
-                    "Wrong render status! Render returned status {}".format(
-                        render_status
-                    )
-                )
+                self.logger.error("Wrong render status", render_status=render_status)
                 self.running_test.becomes_completed()
 
         def render_task_error(e):
-            logger.debug(
-                "render_task_error: task.uuid: {}\n{}".format(render_task.uuid, str(e))
+            self.logger.debug(
+                "render_task_error", task_uuid=render_task.uuid, exc_info=e
             )
             self.running_test.pending_exceptions.append(e)
             self.running_test.becomes_completed()
@@ -125,12 +126,13 @@ class RunningTestCheck(object):
             lambda: self.running_test.rendering_service.render(
                 render_request, render_task_succeeded, render_task_error
             ),
+            self.logger,
         )
         return render_task
 
     def _check_task(self, render_task, tag):
         def check_run():
-            logger.debug("check_run: render_task.uuid: {}".format(render_task.uuid))
+            self.logger.debug("check_run", task_uuid=render_task.uuid)
             self.running_test.eyes.check(
                 self.check_settings,
                 render_task.uuid,
@@ -140,12 +142,14 @@ class RunningTestCheck(object):
             )
 
         check_task = VGTask(
-            "perform check {} {}".format(tag, self.check_settings), check_run
+            "perform check {} {}".format(tag, self.check_settings),
+            check_run,
+            self.logger,
         )
 
         def check_task_completed():
             # type: () -> None
-            logger.debug("check_task_completed: task.uuid: {}".format(check_task.uuid))
+            self.logger.debug("check_task_completed", task_uuid=check_task.uuid)
             if (
                 self.running_test.task_lock
                 and self.running_test.task_lock.uuid == self.uuid
@@ -157,7 +161,7 @@ class RunningTestCheck(object):
                 self.running_test.becomes_tested()
 
         def check_task_error(e):
-            logger.debug("check_task_error: task.uuid: {}".format(check_task.uuid))
+            self.logger.debug("check_task_error", task_uuid=check_task.uuid)
             self.running_test.pending_exceptions.append(e)
             self.running_test.becomes_completed()
 
@@ -171,7 +175,8 @@ class RunningTest(object):
     eyes = attr.ib(hash=False, repr=False)  # type: EyesConnector
     configuration = attr.ib(hash=False, repr=False)  # type: Configuration
     browser_info = attr.ib()  # type: RenderBrowserInfo
-    rendering_service = attr.ib()
+    rendering_service = attr.ib(hash=False, repr=False)  # type: RenderingService
+    logger = attr.ib(hash=False, repr=False)  # type: BoundLogger
 
     tasks_list = attr.ib(init=False, factory=list, hash=False)
     test_uuid = attr.ib(init=False)
@@ -185,6 +190,7 @@ class RunningTest(object):
 
     def __attrs_post_init__(self):
         # type: () -> None
+        self.logger = self.logger.bind(running_test=self)
         self._initialize_vars()
         self._initialize_state_machine()
         self.open()
@@ -245,23 +251,22 @@ class RunningTest(object):
         open_task = VGTask(
             "open {}".format(self.browser_info),
             lambda: self.eyes.open(self.configuration),
+            self.logger,
         )
-        logger.debug("RunningTest %s" % open_task.name)
+        self.logger.debug("RunningTest VGTask created", task_name=open_task.name)
 
         def open_task_succeeded(test_result):
             # type: (Optional[Any]) -> None
-            logger.debug("open_task_succeeded: task.uuid: {}".format(open_task.uuid))
+            self.logger.debug("open_task_succeeded")
             self.watch_open[open_task] = True
             if self.all_tasks_completed(self.watch_open):
                 if self.state == TESTED:
-                    logger.debug("open_task_succeeded: test session was aborted")
+                    self.logger.debug("open_task_succeeded: test session was aborted")
                     return
                 self.becomes_opened()
 
         def open_task_error(e):
-            logger.debug(
-                "open_task_error: task.uuid: {}\n{}".format(open_task.uuid, str(e))
-            )
+            self.logger.debug("open_task_error", exc_info=e)
             self.pending_exceptions.append(e)
             self.becomes_completed()
 
@@ -277,7 +282,7 @@ class RunningTest(object):
         source,  # type: Optional[Text]
     ):
         # type: (...) -> RunningTestCheck
-        logger.debug("RunningTest %s" % check_settings)
+        self.logger.debug("RunningTest.check", check_settings=check_settings)
         running_test_check_task = RunningTestCheck(
             "RunningTestCheck({})".format(check_settings.values.name),
             self,
@@ -299,13 +304,12 @@ class RunningTest(object):
             close_task = VGTask(
                 "close {}".format(self.browser_info),
                 lambda: self.eyes.is_open and self.eyes.close(False),
+                self.logger,
             )
-            logger.debug("RunningTest %s" % close_task.name)
+            self.logger.debug("RunningTest VGTask", task_name=close_task.name)
 
             def close_task_succeeded(test_result):
-                logger.debug(
-                    "close_task_succeeded: task.uuid: {}".format(close_task.uuid)
-                )
+                self.logger.debug("close_task_succeeded")
                 # abort() could add test_result already
                 if self.test_result is None and test_result:
                     self.test_result = test_result
@@ -314,19 +318,13 @@ class RunningTest(object):
 
             def close_task_completed():
                 # type: () -> None
-                logger.debug(
-                    "close_task_completed: task.uuid: {}".format(close_task.uuid)
-                )
+                self.logger.debug("close_task_completed")
                 self.watch_close[close_task] = True
                 if self.all_tasks_completed(self.watch_close):
                     self.becomes_completed()
 
             def close_task_error(e):
-                logger.debug(
-                    "close_task_error: task.uuid: {}\n{}".format(
-                        close_task.uuid, str(e)
-                    )
-                )
+                self.logger.debug("close_task_error", exc_info=e)
                 self.pending_exceptions.append(e)
 
             close_task.on_task_succeeded(close_task_succeeded)
@@ -346,26 +344,26 @@ class RunningTest(object):
                 self.eyes._ensure_running_session()
             return self.eyes.abort()
 
-        abort_task = VGTask("abort {}".format(self.browser_info), ensure_and_abort)
-        logger.debug("RunningTest %s" % abort_task.name)
+        abort_task = VGTask(
+            "abort {}".format(self.browser_info), ensure_and_abort, self.logger
+        )
+        self.logger.debug("RunningTest VGTask created", task_name=abort_task.name)
 
         def abort_task_succeeded(test_result):
-            logger.debug("abort_task_succeeded: task.uuid: {}".format(abort_task.uuid))
+            self.logger.debug("abort_task_succeeded", task_uuid=abort_task.uuid)
             self.test_result = test_result
             if callable(self.on_results):
                 self.on_results(test=self, test_result=test_result)
 
         def abort_task_completed():
             # type: () -> None
-            logger.debug("abort_task_completed: task.uuid: {}".format(abort_task.uuid))
+            self.logger.debug("abort_task_completed", task_uuid=abort_task.uuid)
             self.watch_close[abort_task] = True
             if self.all_tasks_completed(self.watch_close):
                 self.becomes_completed()
 
         def abort_task_error(e):
-            logger.debug(
-                "abort_task_error: task.uuid: {}\n{}".format(abort_task.uuid, str(e))
-            )
+            self.logger.debug("abort_task_error", task_uuid=abort_task.uuid, exc_info=e)
             self.pending_exceptions.append(e)
 
         abort_task.on_task_succeeded(abort_task_succeeded)